commit one before the end

+++ Backend/CateringFinalProject/Entity/Models/TblProducts.cs
@@ -13,7 +13,7 @@ namespace Entity.Models
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public int CategoryId { get; set; }
-        public int? QuantityInStock { get; set; }
+        public double? QuantityInStock { get; set; }
         public int TypeOfMeasurementId { get; set; }
         public int ManagerId { get; set; }
